Minor logic changes. Both unit and integration tests written with 100% coverage. Will add additional assertions if I have time.

+++ src/main/java/management/task/prototal/task_manager/controller/TaskController.java
@@ -1,7 +1,9 @@
 package management.task.prototal.task_manager.controller;
 
+import management.task.prototal.task_manager.exception.InvalidTaskException;
 import management.task.prototal.task_manager.service.TaskService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
@@ -14,30 +16,62 @@ public class TaskController {
     @Autowired
     private TaskService taskService;
 
-    @PostMapping("createTask")
-    public Mono<Task> createTask (@RequestBody Task task) {
-        return taskService.createTask(task);
-    }
-
+    /**
+     * I know this method is superfluous. But it makes it easier for someone
+     * unfamiliar with MongoDB to get the ids for use in the other methods
+     * @return
+     */
     @GetMapping("getAll")
     public Flux<Task> getAllTasks() {
         return taskService.getAllTasks();
     }
 
-    @GetMapping("getById/{id}")
-    public Mono<Task> getTaskById(@PathVariable String id) {
-        return taskService.getTaskById(id);
+    @PostMapping("createTask")
+    public Mono<Task> createTask (@RequestBody Task task) {
+        try {
+            if (task == null || task.getDescription() == null || task.getTitle() == null) {
+                throw new InvalidTaskException("Task or task properties cannot be null");
+            }
+            return taskService.createTask(task);
+        } catch (InvalidTaskException e) {
+            return Mono.error(e);
+        }
     }
 
-    // This seems to only update the id. Change this. Dummy method. Will change.
-    @PutMapping("/{id}")
+    @GetMapping("getById/{id}")
+    public Mono<Task> getTaskById(@PathVariable String id) {
+        try {
+            if (id == null) {
+                throw new InvalidTaskException("ID is null");
+            }
+            return taskService.getTaskById(id);
+        } catch (InvalidTaskException e) {
+            return Mono.error(e);
+        }
+    }
+
+    @PutMapping("/update/{id}")
     public Mono<Task> updateTask(@PathVariable String id, @RequestBody Task task) {
-        task.setId(id);
-        return taskService.updateTask(task);
+        try {
+            if (task == null || task.getDescription() == null || task.getTitle() == null) {
+                throw new InvalidTaskException("Task or task properties cannot be null");
+            }
+            task.setId(id);
+            return taskService.updateTask(task);
+        } catch (InvalidTaskException e) {
+            return Mono.error(e);
+        }
     }
 
     @DeleteMapping("deleteTask/{id}")
     public Mono<Boolean> deleteTask(@PathVariable String id) {
-        return taskService.deleteTask(id);
+        try {
+            if (id == null || id.equals("null")) {
+                throw new InvalidTaskException("Id is null");
+            }
+            return taskService.deleteTask(id);
+        } catch (InvalidTaskException e) {
+            return Mono.error(e);
+        }
     }
 }
